refactor `patient_canonical_identifier` to patient module for better reuse.
