+++ carl/logic/copd.py
@@ -1,12 +1,10 @@
 """COPD module"""
-from flask import current_app, has_app_context
-import requests
+from flask import current_app
 
-from carl.config import FHIR_SERVER_URL
 from carl.modules.coding import Coding
 from carl.modules.codeableconcept import CodeableConcept
 from carl.modules.condition import Condition
-from carl.modules.patient import Patient, patient_has
+from carl.modules.patient import Patient, patient_canonical_identifier, patient_has
 from carl.modules.resource import delete_resource, persist_resource
 from carl.modules.valueset import valueset_codings
 
@@ -34,26 +32,6 @@ CNICS_COPD_medication_coding = Coding(
     code="CNICS.COPDMED2022.03.001",
     display="COPD PRO group member with qualifying MedicationRequest",
 )
-
-CNICS_IDENTIFIER_SYSTEM = "https://cnics.cirg.washington.edu/site-patient-id/"
-
-
-def patient_canonical_identifier(patient_id, site_code):
-    """Return system|value identifier if patient has one for preferred system"""
-    url = f"{FHIR_SERVER_URL}Patient/{patient_id}"
-    response = requests.get(url, timeout=30)
-    if has_app_context():
-        current_app.logger.debug(f"HAPI GET: {response.url}")
-    response.raise_for_status()
-
-    match = [
-        f"{identifier['system']}|{identifier['value']}"
-        for identifier in response.json().get("identifier", [])
-        if identifier["system"] == CNICS_IDENTIFIER_SYSTEM + site_code
-    ]
-
-    if match:
-        return match[0]
 
 
 def process_4_COPD_conditions(patient_id, site_code):
